Update Java client failure handling
The current description does not handle the case where a Java client connects to a member successfully, but then the member becomes unresponsive without closing the connection. Previous versions sounded like this was able to be handled by invocation timeout, but discussions with Engineering indicate that only the Client Failure Detectors can actually handle this scenario.